+++ app/api/paragraphExtraction/infrastructure/PXCreateParagraphsFactory.ts
@@ -3,14 +3,14 @@ import { getConnection } from 'api/common.v2/database/getConnectionForCurrentTen
 
 import { PXCreateParagraphs } from '../application/PXCreateParagraphs';
 import { MongoPXExtractorsDataSource } from './MongoPXExtractorsDataSource';
-import { MongoPXExtractionsDataSource } from './MongoPXExtractionsDataSource';
+import { MongoPXEntitiesStatusDataSource } from './MongoPXEntitiesStatusDataSource';
 
 export class PXCreateParagraphsFactory {
   static createDefault() {
     const db = getConnection();
     const transactionManager = DefaultTransactionManager();
     const extractorsDS = new MongoPXExtractorsDataSource(db, transactionManager);
-    const extractionsDS = new MongoPXExtractionsDataSource(db, transactionManager);
+    const extractionsDS = new MongoPXEntitiesStatusDataSource(db, transactionManager);
 
     return new PXCreateParagraphs({
       extractorsDS,
