+++ app/api/paragraphExtraction/infrastructure/PXExtractParagraphsFromEntitiesFactory.ts
@@ -3,14 +3,14 @@ import { DefaultTransactionManager } from 'api/common.v2/database/data_source_de
 import { DefaultDispatcher } from 'api/queue.v2/configuration/factories';
 
 import { PXExtractParagraphsFromEntities } from '../application/PXExtractParagraphFromEntities';
-import { MongoPXExtractionsDataSource } from './MongoPXExtractionsDataSource';
+import { MongoPXEntitiesStatusDataSource } from './MongoPXEntitiesStatusDataSource';
 
 export class PXExtractParagraphsFromEntitiesFactory {
   static async createDefault(tenantName: string) {
     const db = getConnection();
     const transactionManager = DefaultTransactionManager();
 
-    const extractionsDS = new MongoPXExtractionsDataSource(db, transactionManager);
+    const extractionsDS = new MongoPXEntitiesStatusDataSource(db, transactionManager);
 
     const dispatcher = await DefaultDispatcher(tenantName, { lockWindow: 1000 * 60 });
