+++ app/api/paragraphExtraction/infrastructure/PXParagraphsResultListener.ts
@@ -9,7 +9,7 @@ import { PXExtractionService } from '../domain/PXExtractionService';
 import { PXExtractionServiceFactory } from './PXExtractionServiceFactory';
 import { PXCreateParagraphsFactory } from './PXCreateParagraphsFactory';
 import { PXExtractionKey } from '../domain/PXExtractionKey';
-import { MongoPXExtractionsDataSource } from './MongoPXExtractionsDataSource';
+import { MongoPXEntitiesStatusDataSource } from './MongoPXEntitiesStatusDataSource';
 
 type ResultMessage = {
   key: string;
@@ -54,7 +54,7 @@ export class PXParagraphsResultListener {
         await this.setCurrentUser(extractionKey.userId);
         await this.getUseCase().execute(result);
       } catch (e) {
-        const extractionsDS = new MongoPXExtractionsDataSource(
+        const extractionsDS = new MongoPXEntitiesStatusDataSource(
           getConnection(),
           DefaultTransactionManager()
         );
